ci: Add GitHub Actions for container build and deploy
This commit introduces a new GitHub Actions workflow to automate
the build and deployment of Docker containers. The workflow is
configured to trigger on all pushes to the master branch and on
manual dispatch.

Key features include:
- A new GitHub Actions workflow file, `containers.yml`, that
  defines the build and deploy logic for Leap AI container
  images.
- Utilizes `redhat-actions/podman-login` for authentication with
  the GitHub Container Registry (GHCR).
- Employs `aardsoft/ci-actions/kiwi-build` for building the
  containers specified in the `leap-rocm-base` directory.
- Includes additional repository links for obtaining required
  RPM packages necessary for building Leap AI container images.

Additionally, a new `ollama.kiwi` file in the `leap-rocm-base`
directory defines the container image structure, pointing to the
base Leap image and listing necessary packages such as `libpci3` and
various ROCm libraries. This ensures the container is equipped
with tools for running ROCm applications.